+++ sources/packages/web/src/router/AESTRoutes.ts
@@ -17,7 +17,7 @@ import Designation from "@/views/aest/institution/Designation.vue";
 import Restrictions from "@/views/aest/institution/Restrictions.vue";
 import InstitutionNotes from "@/views/aest/institution/InstitutionNotes.vue";
 import ApplicationDetails from "@/views/aest/ApplicationDetails.vue";
-import StudentApplication from "@/views/aest/StudentApplication.vue";
+import StudentApplicationView from "@/views/aest/StudentApplicationView.vue";
 import AESTHomeSideBar from "@/components/layouts/aest/AESTHomeSideBar.vue";
 import StudentNotes from "@/views/aest/student/StudentNotes.vue";
 import StudentRestrictions from "@/views/aest/student/StudentRestrictions.vue";
@@ -139,7 +139,7 @@ export const aestRoutes: Array<RouteRecordRaw> = [
             path: "",
             name: AESTRoutesConst.APPLICATION_DETAILS,
             props: true,
-            component: StudentApplication,
+            component: StudentApplicationView,
             meta: {
               clientType: ClientIdType.AEST,
             },
